+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 - [usd#2625](https://github.com/Autodesk/arnold-usd/issues/2625) - Add turntable command line tool generating turntable scenes from a usd asset.
 - [usd#1286](https://github.com/Autodesk/arnold-usd/issues/1286) - Support UV RenderVars
+- [usd#2687](https://github.com/Autodesk/arnold-usd/issues/2687) - Support double-precision primvars
 
 ### Bug Fixes